Assert same for all files in the xyz.docx document.
Comparing the unzip files is necessary. Two word documents are only the same if they are generated exactly at the same timestamp. (This might be caused by the zip algorithm or some other word file config that can not be manipulated by the PhpWord library.) Therefore, the unzipped content gets compared.